Added the newest version of forth.py
I rewrote the forth.py from scratch for the fourth time, adding in a simple parser. It could still be further divided to meet best practices, but I did't have enough time. I still have to rewrite the REPL from scratch to make it work. It also only includes primitives as of right now, but that's easy enough to fix.